+++ app/prep.f90
@@ -87,6 +87,7 @@ use M_list,      only : insert, locate, replace, remove                      ! B
    character(len=:),allocatable,save    :: G_extract_start
    character(len=:),allocatable,save    :: G_extract_stop
    logical,save                         :: G_extract=.false.
+   logical,save                         :: G_extract_auto=.false.
    logical,save                         :: G_extract_flag=.false.
 
    character(len=:),allocatable         :: keywords(:)
@@ -2042,6 +2043,7 @@ help_text=[ CHARACTER(LEN=128) :: &
 '         [--width n]                                                            ',&
 '         [-d ignore|remove|blank]                                               ',&
 '         [--comment default|doxygen|ford|none]                                  ',&
+'         [--type FILE_TYPE | --start START_STRING --stop STOP_STRING]           ',&
 '         [--ident]                                                              ',&
 '         [--version]                                                            ',&
 '         [--help]                                                               ',&
@@ -2139,19 +2141,37 @@ help_text=[ CHARACTER(LEN=128) :: &
 '                        ASCII Decimal Equivalent (Common values are 37=%        ',&
 '                        42=* 35=# 36=$ 64=@). If the value is not numeric       ',&
 '                        it is assumed to be a literal character.                ',&
-'                                                                                ',&
-'   --help           Display documentation and exit.                             ',&
-'                                                                                ',&
-'   --verbose        All commands on a $SYSTEM directive are echoed              ',&
-'                    to stderr with a + prefix. Text following the               ',&
-'                    string "@(#)" is printed to stderr similar to               ',&
-'                    the Unix command what(1) but is otherwise                   ',&
-'                    treated as other text input.                                ',&
-'                                                                                ',&
 '   --noenv          The $IFDEF and $IFNDEF directives test for an               ',&
 '                    internal prep(1) variable and then an                       ',&
 '                    environment variable by default. This option                ',&
 '                    turns off testing for environment variables.                ',&
+'                                                                                ',&
+'   --type FILETYPE  this flag indicates to skip input lines until after a       ',&
+'                    specific start string is encountered and to stop once a     ',&
+'                    specific end string is found left-justifed on lines by      ',&
+'                    themselves.                                                 ',&
+'                                                                                ',&
+'                       FILETYPE  START_STRING   STOP_STRING                     ',&
+'                          md      ```fortran     ```                            ',&
+'                        html      <xmp>          </xmp>                         ',&
+'                        auto                                                    ',&
+'                        none                                                    ',&
+'                                                                                ',&
+'                    The special type "auto" may be specified, in which case     ',&
+'                    files will be processed according to their file suffix.     ',&
+'                    This allows for easily extracting code from common          ',&
+'                    document formats. This is particularly useful with extended ',&
+'                    markdown formats, allowing for code source to be easily     ',&
+'                    documented and for tests in documents to be able to be      ',&
+'                    extraccted and tested. "auto" switches processing mode      ',&
+'                    depending on input file suffix, treating supported file     ',&
+'                    prefixs ("md","html") appropriately.                        ',&
+'                                                                                ',&
+'    --start STRING  Same as --type except along with --stop allows for custom   ',&
+'                    strings to be specified.                                    ',&
+'                                                                                ',&
+'    --stop STRING   Same as --type except along with --start allows for custom  ',&
+'                    strings to be specified.                                    ',&
 '                                                                                ',&
 '   --system         Allow system commands on $SYSTEM directives to              ',&
 '                    be executed.                                                ',&
@@ -2184,12 +2204,20 @@ help_text=[ CHARACTER(LEN=128) :: &
 '                           "debug" code by many Fortran compilers when          ',&
 '                           compiling fixed-format Fortran source.               ',&
 '                                                                                ',&
-'   --version        Display version and exit                                    ',&
-'                                                                                ',&
 '   --width n        Maximum line length of the output file. The default is 1024.',&
 '                    Typically used to trim fixed-format FORTRAN code that       ',&
 '                    contains comments or "ident" labels past column 72          ',&
 '                    when compiling fixed-format Fortran code.                   ',&
+'                                                                                ',&
+'   --verbose        All commands on a $SYSTEM directive are echoed              ',&
+'                    to stderr with a + prefix. Text following the               ',&
+'                    string "@(#)" is printed to stderr similar to               ',&
+'                    the Unix command what(1) but is otherwise                   ',&
+'                    treated as other text input.                                ',&
+'                                                                                ',&
+'   --version        Display version and exit                                    ',&
+'                                                                                ',&
+'   --help           Display documentation and exit.                             ',&
 '                                                                                ',&
 '   DIRECTIVES                                                                   ',&
 '                                                                                ',&
@@ -2611,7 +2639,8 @@ help_text=[ CHARACTER(LEN=128) :: &
 '@(#)DESCRIPTION:    Fortran Pre-processor>',&
 !'@(#)VERSION:        4.0.0: 20170502>',&
 !'@(#)VERSION:        5.0.0: 20201219>',&
-'@(#)VERSION:        6.0.0: 20210613>',&
+!'@(#)VERSION:        6.0.0: 20210613>',&
+'@(#)VERSION:        6.0.1: 20220311>',&
 '@(#)AUTHOR:         John S. Urban>',&
 '@(#)HOME PAGE       https://github.com/urbanjost/prep.git/>',&
 '']
@@ -3010,6 +3039,20 @@ end subroutine get_os_type
 !===================================================================================================================================
 !()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()!
 !===================================================================================================================================
+pure function ends_in(string) result(ending)
+character(*), intent(in)  :: string
+character(:), allocatable :: ending
+integer                   :: n1
+   n1=index(string,'.',back=.true.)
+   if (n1 < 1 .or. n1.eq.len(string) ) then
+       ending=''
+   else
+       ending=string(n1+1:)
+   endif
+end function ends_in
+!===================================================================================================================================
+!()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()!
+!===================================================================================================================================
 end module M_fpp
 !===================================================================================================================================
 !()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()!
@@ -3105,9 +3148,19 @@ logical                       :: isscratch
       ! flaw is HTML is not case sensitive
       G_extract_start='<xmp>'
       G_extract_stop='</xmp>'
+   case('auto')
+      G_extract_start=''
+      G_extract_stop=''
+      G_extract_auto=.true.
+      G_extract=.true.
+   case('none')
+      G_extract_start=''
+      G_extract_stop=''
+      G_extract_auto=.false.
+      G_extract=.false.
    case default
-      G_extract_start=sget('prep_start')
-      G_extract_stop=sget('prep_stop')
+      G_extract_start=trim(sget('prep_start'))
+      G_extract_stop=trim(sget('prep_stop'))
    end select
    if(G_extract_start.ne.''.or.G_extract_stop.ne.'')G_extract=.true.
 
@@ -3115,6 +3168,7 @@ logical                       :: isscratch
    call defines()                                          ! define named variables declared on the command line
    call includes()                                         ! define include directories supplies on command line
    call opens()                                            ! convert input filenames into $include directives
+   call auto()
 
    READLINE: do                                            ! read loop to read input file
       read(G_file_dictionary(G_iocount)%unit_number,'(a)',end=7) line
@@ -3181,12 +3235,33 @@ logical                       :: isscratch
       endif
 
       if(G_iocount.lt.1)exit
+      call auto() ! if in auto mode determine strings for new file
+
    enddo READLINE
 
    if (G_nestl.ne.0) then                                           ! check to make sure all if blocks are closed
       call stop_prep('*prep* ERROR(067) - $IF BLOCK NOT CLOSED.')
    endif
    call print_comment_block()
+
+   contains
+
+subroutine auto()
+   if(G_extract_auto)then
+      select case(ends_in(G_file_dictionary(G_iocount)%filename) )
+      case('md','.md')
+         G_extract_start='```fortran'
+         G_extract_stop='```'
+      case('html','.html','htm','.htm')
+         G_extract_start='<xmp>'
+         G_extract_stop='</xmp>'
+      case default
+         G_extract_start=trim(sget('prep_start'))
+         G_extract_stop=trim(sget('prep_stop'))
+      end select
+   endif
+end subroutine auto
+
 end program prep
 !===================================================================================================================================
 !()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()!